nodejs: enhance e2e testing between nodejs and .NET- test case cosmetics- commission at least 2 test cases
