+++ src/Clients/nodejs/test/unit/core-e2e/helpers/line-emitter.ts
